requirements.txt
flask
rembg[cpu]
Pillow
python-dotenv
gunicorn
werkzeug
